requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.3.0
holidays==0.83
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.2.1
colorama==0.4.6
streamlit>=1.28.0
pandas>=1.5.0
mysql-connector-python>=8.0.0
comm==0.2.2
contourpy==1.3.2
curl_cffi==0.11.4
cycler==0.12.1
decorator==5.2.1
distlib==0.4.0
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.1
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kaleido==1.0.0
kiwisolver==1.4.8
logging==0.4.9.6
logistro==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
multitasking==0.0.11
narwhals==1.42.1
nbformat==5.10.4
numpy==2.2.6
openpyxl==3.1.5
optional==0.0.1
orjson==3.10.18
packaging==25.0
pandas==2.3.1
parso==0.8.4
path==17.1.0
peewee==3.18.2
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
prompt_toolkit==3.0.51
protobuf==6.31.1
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.25.1
seaborn==0.13.2
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
smartsheet-python-sdk==3.0.5
smmap==5.0.2
soupsieve==2.7
stack-data==0.6.3
streamlit==1.46.1
streamlit-aggrid==1.1.6
streamlit-javascript>=0.1.5
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
typing==3.7.4.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.32.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14

yfinance==0.2.65

